First batch of staypuft models
This includes the new custom models for staypuft and the join models
for the has_many :through relations. Not yet included are the
join models which connect to native foreman types (HostGroup, etc.)
or any extensions to foreman models -- that will be a later commit.

Also, the scaffolding for tests is included for now, but it will
be removed when we add the model rspec tests.

on edit/update:
1) removed "OpenStack" from model classnames
2) removed empty test cases
3) modified association validations to validate the object rather than the ID
4) added fk constraints using foreigner