Added networking to the Layout model
FIXME: not merge-worthy yet, but I wanted to get this
work-in-progress out there before I'm offline for a couple
days.

Refactor Layout to have a unique key of (name, networking)
in order to treat, for example. "HA Nova" and "HA Neutron"
as separate layouts. The UI will still expose this at
deployment time as a choice of a named layout and a
type of networking.