don't display the ugly auto-generated deployment name: make user choose a new one
We auto-generate an ugly name since name is required in the Deployment model
and the wizard requires the object be created prior to showing the 'new' form.
This commit adds a specific "uninitialized deployment" prefix to the auto-generated
name, and the wizard clears it out of the form, requiring the user to enter
a real name on the first step of the wizard.