add HA layout to staypuft
FOr this to work you need to update astapor and openstack-puppet-modules
to the latest 'master' commit. The astapor HA stuff is still being
actively developed, so this is probably still unstable.

YOu need to load the new puppetclasses too, and I don't know if you can simply
re-run the puppet module import or if you have to drop db and re-run everything.

In any case this also requires re-running rake db:seed after getting the new
astapor stuff and this stuff.