Reverting to old version of host ready method
The puppet modules can sometimes fail then in subsequent run
after a new host is alive, it can go to active state.  The
modules need update to change this behaviour.  For now we
are accounting for this behaviour in our DynFlow tasks, where
we are only check to see if progress has been made, before
scheduling the next host.