Reset host token during build task
Foreman sets the expire provisioning token on a host machine when host
machine build flag is set, not when the host is actually provisioned.
In Staypuft we delay the provisioning of some machines until later in
the orchestration sequence.  This may result in token expiry.  This
patch resets the host expire token during the Staypuft host build task.