Add deployment form_step for condition validations
This patch adds an extra field form_step to the deployment model.  This
is to be used for conditional validations.  Validations can now take
place in a aggregational model, where each step in the deployment
creation form has it's own validations.  This method is taken directly
from the wicked rubygem wiki.