Get params in sync with latest astapor modules.
For node types that only had one service, this PR moves to the
param-handling style we use for the HA controller, where we don't
hard-code the param list, since they're already broken down to the
service/puppetclass level.

For neutron networker and the non-HA controllers, we've got to stick
with the old system, so this updates the param list to reflect the
latest astapor codebase.