rabbit/qpid UI switch
Add selection of AMQP Provider on first page of wizard
(RabbitMQ vs. Qpid). Sets the param on the deployment, and
then iterates over the hostgroups and sets the amqp_server param
(soon changing to amqp_provider) on those puppetclasses that
define it (much as we do for nova vs. neutron)

db:migrate is required with this PR