bz 1118502: set param type explicitly for erb params
for data seeding, we use the default value class to determine
the proper foreman/puppet param type. This doesn't work for the
dynamic params, since the object type is always a string -- i.e.
an erb snippet,but it doesn't always *evaluate* to a string,
as it could be an array or a hash.