hack to get deployment.ips.controller_ips to work w/ multi-nics
host.ip doesn't return a predictable value when there is more than
one network interface. Once we explicitly support multiple subnets
in staypuft w/ the phase 3 development, we'll need something
less hackish than this -- "give me the IP for this host on $subnet"
It could be that we already have a reliable way to do this without
this DNS hack, in which case we should do that instead.

In any case, this seems to do the right thing for current multi-nic
deployments