calculate nova network size
This gets rid of the default '256' nova network size, which required larger
fixed tenant ranges when using vlans. Now, network size is calculated
from the fixed tenant range and the number of networks, to make sure that
the networks all fit within the same address space.

This also includes validation to make sure that the fixed range is
large enough to allow for a minimum network size of 4.