Fencing: refactor controller concern to use mock model
Refactor the fencing controller concern to use the mock Fencing model
on Host. With this approach, the request params no longer need to be
mofified manually, and allows the usage of validations on the mock
model. The patch also adds a validation of the presence of the
fencing_type value.

partially fixes BZ #1144649