for discovery environment, ignore power management (bmc, etc)
We hit an issue where configuring fencing on the host pre-deployment caused
deploy reboot to fail. As it turns out, simply having the bmc nic's
username and password field filled in causes foreman to want to control
the host via bmc. This may be fine post-deployment (although we
probably have lingering bmc proxy problems for that to really work),
but for initial deployment, this breaks and I don't know that we
want to boot fencing-enabled hosts differently than others, even
if it does work -- so with this, discovery env hosts are rebooted
via the discovery image proxy regardless of whether they're
configured for fencing post-deployment