BZ #1177740: Fix bond mode selection for multiple hosts
https://bugzilla.redhat.com/show_bug.cgi?id=1177740

This will also fix for the other operations (adding additional
interfaces, deleting interfaces, etc.). The list of hosts was not being
parsed properly. Therefore, only one of the hosts was being operated
upon. When the user went back into the settings, it would reset to
whatever setting the first host has, which may not have been what was
selected.