BZ#1173634 - deployment stuck
https://bugzilla.redhat.com/show_bug.cgi?id=1173634

This fixes another of possible causes of deployment getting
stuck. After the host boots up, we attempt to run puppet on it as soon
as we detect that port 22 is open, but that doesn't mean that host is
really finished booting and ready to run puppet. We now have a 60
second grace period between detecting ssh port is open and actually
triggering a puppet run.