BZ#1180322 - registration puppet run reports errors
https://bugzilla.redhat.com/show_bug.cgi?id=1180322

Even though the client registration Puppet run doesn't apply any
changes, it still attempts to compile the full host's Puppet catalog and
may cause errors. Switching it to an empty environment will exclude the
Puppet classes from the YAML description of the node and cause the
catalog to be empty and without errors.